Don’t error if organisation name is unchanged
If you submit the rename organisation form without making any changes
you will get an error saying that the name is currently in use. This is
true because it’s being used by the current organisation.

However your intention is probably not to actually change anything, so
we can just redirect back to the settings page.

This is the same thing we do when renaming services:
https://github.com/alphagov/notifications-admin/blob/60f5b7490412406695d535060e13bc6184d901df/app/main/views/service_settings.py#L99-L100